[chore] reformat FormWizard for reliability
We want this library to be demoable.

Changes:
- Renames the folder.
- Exports components in sensible fashion.
- Shares validate method across components.
- Add two render prop options (component or children.
- Exposes the submit handler.
