+++ www/src/pages/store/components/ItemDetailsFieldSet.js
@@ -1,11 +1,10 @@
 import React from 'react'
 
 import {
-  FormContext,
   EmailField,
   TextField,
   FileField,
-} from '../../../shared/form-wizard/FormWizard'
+} from '../../../shared/FormWizard'
 
 class ItemDetailsFieldSet extends React.Component {
   constructor(props){
@@ -16,50 +15,38 @@ class ItemDetailsFieldSet extends React.Component {
 
   render() {
     return (
-      <FormContext.Consumer>
-        {(context) => {
-          return (
-            <React.Fragment>
-              <FileField
-                name="item_thumbnail_image"
-                label="Thumbnail preview"
-                placeholder='Thumbnail image of item'
-                updateFormDataHandler={(e) => context.updateFormData(e)}
-              />
+      <React.Fragment>
+        <FileField
+          name="item_thumbnail_image"
+          label="Thumbnail preview"
+          placeholder='Thumbnail image of item'
+        />
 
-              <TextField
-                name="item_name"
-                label="What is the name of your theme?"
-                placeholder='Name'
-                updateFormDataHandler={(e) => context.updateFormData(e)}
-              />
+        <TextField
+          name="item_name"
+          label="What is the name of your theme?"
+          placeholder='Name'
+        />
 
-              <TextField
-                name="item_description"
-                label="What description do you want customers to see?"
-                placeholder='Description'
-                updateFormDataHandler={(e) => context.updateFormData(e)}
-              />
+        <TextField
+          name="item_description"
+          label="What description do you want customers to see?"
+          placeholder='Description'
+        />
 
-              <TextField
-                name="item_demo_url"
-                label="Where is the demo url?"
-                placeholder='Demo url'
-                updateFormDataHandler={(e) => context.updateFormData(e)}
-              />
+        <TextField
+          name="item_demo_url"
+          label="Where is the demo url?"
+          placeholder='Demo url'
+        />
 
-              <EmailField
-                name="item_support_email"
-                label="How do customers get support for your theme?"
-                placeholder="Email address"
-                updateFormDataHandler={(e) => context.updateFormData(e)}
-              />
-            </React.Fragment>
-          )
-        }}
-      </FormContext.Consumer>
+        <EmailField
+          name="item_support_email"
+          label="How do customers get support for your theme?"
+          placeholder="Email address"
+        />
+      </React.Fragment>
     )
   }
-
 }
  export { ItemDetailsFieldSet }
