+++ www/src/pages/store/components/ItemPriceDataFieldSet.js
@@ -1,9 +1,8 @@
 import React from 'react'
 
 import {
-  FormContext,
   NumberField,
-} from '../../../shared/form-wizard/FormWizard'
+} from '../../../shared/FormWizard'
 
 class ItemPriceDataFieldSet extends React.Component {
   constructor(props){
@@ -14,28 +13,19 @@ class ItemPriceDataFieldSet extends React.Component {
 
   render() {
     return (
-      <FormContext.Consumer>
-        {(context) => {
-          return (
-            <React.Fragment>
-              <NumberField
-                name="item_standard_price"
-                label="What standard price do you want to set?"
-                placeholder='Standard Price'
-                updateFormDataHandler={(e) => context.updateFormData(e)}
-              />
+      <React.Fragment>
+        <NumberField
+          name="item_standard_price"
+          label="What standard price do you want to set?"
+          placeholder='Standard Price'
+        />
 
-              <NumberField
-                name="item_commercial_price"
-                label="What commercial price do you want to set?"
-                placeholder='Commercial Price'
-                updateFormDataHandler={(e) => context.updateFormData(e)}
-              />
-            </React.Fragment>
-
-          )
-        }}
-      </FormContext.Consumer>
+        <NumberField
+          name="item_commercial_price"
+          label="What commercial price do you want to set?"
+          placeholder='Commercial Price'
+        />
+      </React.Fragment>
     )
   }
