+++ www/src/shared/FormWizard/fields/index.js
@@ -1,0 +1,12 @@
+import { EmailField } from './EmailField'
+import { FileField } from './FileField'
+import { NumberField } from './NumberField'
+import { TextField } from './TextField'
+
+
+export {
+  EmailField,
+  FileField,
+  NumberField,
+  TextField,
+}
