+++ www/src/shared/FormWizard/stepper/index.js
@@ -1,0 +1,7 @@
+import { Step } from './Step'
+import { StepList } from './StepList'
+
+export {
+  StepList,
+  Step,
+}
